requirements.txt
streamlit==1.31.0
google-cloud-aiplatform==1.129.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
requests
